requirements.txt
Flask>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
werkzeug>=3.0.0
openpyxl>=3.1.0
openai>=1.0.0
requests>=2.31.0
livereload>=2.6.3
python-dotenv>=1.0.0
